Allow to specify files to be ignored by cppcheck in kde-ci.yml
Ignoring files can be needed e.g. for large auto-generated files that
take way too long to process for cppcheck. So far this had to be done by
setting `cppcheck-arguments` explicitly, having to repeat the arguments
from the global config file. Now we have a separate config entry,
`cppceck-ignore-files`, which can hold a list of files to be excluded.