reuse: don't put job in the 'test' stage
Having the reuse job in the test stage means it only runs once the build
jobs pass. This is unnecessarily reducing parallelism. Keeping it in the
build stage will let it run in parallel to actual build jobs.

Using different stages is mainly useful when you need previous stages to
finish successfully (for example if we had separate unit test jobs, the
build needs to succeed first).