Put reuse job in 'build' stage
In 39c1a135 I removed "stage: test" from the reuse job, but that's actually the
default stage, so my change didn't do anything. Set it to the build stage
explicitly instead.

This will let it run in parallel to actual build jobs instead of waiting for
the build jobs to finish.