Introduce hierarchical platform flavor configuration
This is motivated by some of the Qt 6 CI builds needing different
dependency or unit test options, but we are seeing similar needs in
the static or custom build configurations as well, so better solve
this sufficiently generically.

This extends the simple OS-only platform string to a hierarchical
platform specifier, for backward compatibility we assume the first
element is still the OS. The main feature however is that this is
partially matched against configurations, ie. "Linux" applies to
e.g. "Linux/Qt6" and "Linux/Qt5/static".