Provide isolated XDG_RUNTIME_DIR for every build
In kwin, we seem to experience an issue where some tests fail when
running in FreeBSD CI. The common denominator is that kwin fails to
create a wayland socket with a concrete name.

As far as I can tell, we started noticing it after adding FreeBSD Qt 6
pipeline. My guess is that two parallel builds conflict with each other.

With this change, run-ci-build.py script will try to allocate an
isolated runtime dir for every ci build.