Ignore libraries/qca when resolving dependencies
qca is treated differently than other KDE dependencies and is shipped as part of the base image.

That means that it isn't supposed to be specified as a dep in .kde-ci.yml and doing so will cause the build to fail because it isn't found.

However, specifying qca in .kde-ci.yml would be useful because it is needed to generate correct dependency data for kdesrc-build

To make this possible ignore qca when resolving dependencies