Use Android toolchain file from ECM CI artifacts
This works for all packages depending on ECM. In particular, it doesn't
work for ECM itself though. We therefore expose the toolchain path as a
Gitlab variable that ECM can then set to the path in its own source dir.
If we have repos that cannot depend on ECM for whatever reason, setting
that to the path in the image is also an option.