Archive everything needed for creating an Android app bundle
We archive everything needed from the arm64 job, and we archive only the
architecture-specific bits (i.e. the compiled libraries) for arm32 and
x86-64.

Removing artifacts:expose_as since it is incompatible with glob patterns
(which we use in artifacts:paths for arm64). It anyway made very little
sense to expose those job artifacts in the merge request UI.